Move key related test cases to own file
